chore: separate platform files, DRY, cleanup
